+++ src/export.ts
@@ -1,6 +1,6 @@
 import Button from './lib/button/button';
 import Chip from './lib/chip';
-import DropDownMenu from './lib/menu/dropdown_menu';
+import DropDownMenu, { DropDownMenuItemType as DropDownMenuItemTypeTemp } from './lib/menu/dropdown_menu';
 import FileInput from './lib/form/file_input';
 import FormDialog, { FormDialogButton as FormDialogButtonType } from './lib/dialog/formDialog';
 import Grid from './lib/layout/grid';
@@ -14,6 +14,7 @@ import Modal from './lib/dialog/modal';
 import PDFViewer from './lib/pdf_viewer/pdf_viewer';
 import PeriodSelector, { PeriodOptions } from './lib/date_time/period_selector';
 import Table, { TableSortDirection as TableSortDirectionType } from './lib/table/table';
+import TabSwitch from './lib/layout/tab_switch';
 import Textarea from './lib/form/textarea';
 import TextareaSet from './lib/form/textarea_set';
 import TimePicker from './lib/form/time_picker';
@@ -41,6 +42,7 @@ export {
     PeriodOptions,
     PeriodSelector,
     Table,
+    TabSwitch,
     Textarea,
     TextareaSet,
     TimePicker,
@@ -55,6 +57,7 @@ export type ListItem = ListItemType
 export type FormDialogButton = FormDialogButtonType
 export type TableSortDirection = TableSortDirectionType
 export type RadioButtonType = RadioButtonTempType
+export type DropDownMenuItemType = DropDownMenuItemTypeTemp
 
 
 export {