+++ src/view/main/main.tsx
@@ -20,6 +20,7 @@ import { PDFViewerFullScreenCreator, pdfViewerFullScreenOptions } from '../pdf_v
 import { FormDialogCreator, formDialogOptions } from '../dialog/form_dialog';
 import { TableCreator, tableOptions } from '../table/table';
 import { RadioGroupCreator, radioGroupOptions } from '../form/radio_group';
+import { TabSwitchCreator, tabSwitchOptions } from '../layout/tab_switch';
 
 
 export default function MainPage(props: {}): JSX.Element {
@@ -43,6 +44,7 @@ export default function MainPage(props: {}): JSX.Element {
         "Form Dialog": { title: "Form Dialog", data: formDialogOptions(), creator: FormDialogCreator },
         "Table": { title: "Table", data: tableOptions(), creator: TableCreator },
         "Radio Group": { title: "Radio Group", data: radioGroupOptions(), creator: RadioGroupCreator },
+        "Tab Switch": { title: "Tab Switch", data: tabSwitchOptions(), creator: TabSwitchCreator },
     }
 
     let content: React.ReactNode = null;